requirements.txt
langchain-openai
langchain
langchain_community
langchain_ollama
python-dotenv
streamlit
langchain_core
ipykernel
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
pypdf 
bs4
arxiv
langchain_groq
duckduckgo-search
youtube_transcript_api
unstructured
pytube
validators==0.28.1
mysql-connector-python
SQLAlchemy
numexpr